+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # NEXT UP
 
 - Some actions would not pick up the correct ability score.
+- Compendium folder creation is enforced for Class, Class Features, Race, and Racial Traits.
+
 # 4.0.2
 
 - Resoruce handling dialogue now has a Remove option, there is a setting to choose if you want to show/hide this by default on new characters. This is now set to hide as default.